+++ app/Http/Controllers/GalleryController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
+use App\Concerns\FormatsWhatsAppNumber;
 use App\Models\ArtistProfile;
 use App\Models\PortfolioItem;
 use App\Models\Setting;
@@ -13,6 +14,7 @@ use Illuminate\View\View;
 
 class GalleryController extends Controller
 {
+    use FormatsWhatsAppNumber;
     public function index(): View
     {
         $portfolioItems = PortfolioItem::with(['category', 'artist'])
@@ -81,24 +83,5 @@ class GalleryController extends Controller
             'portfolioItems' => $portfolioItems,
             'whatsappNumber' => $whatsappNumber,
         ]);
-    }
-
-    protected function formatWhatsAppNumber(string $number): string
-    {
-        $number = preg_replace('/[^0-9]/', '', $number);
-
-        if (str_starts_with($number, '08')) {
-            $number = '62' . substr($number, 1);
-        }
-
-        if (str_starts_with($number, '628')) {
-            return $number;
-        }
-
-        if (str_starts_with($number, '62')) {
-            return $number;
-        }
-
-        return '62' . $number;
     }
 }