+++ app/Http/Controllers/ShopController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
+use App\Concerns\FormatsWhatsAppNumber;
 use App\Models\Category;
 use App\Models\Product;
 use App\Models\Setting;
@@ -13,6 +14,7 @@ use Illuminate\Support\Str;
 
 class ShopController extends Controller
 {
+    use FormatsWhatsAppNumber;
     public function __construct(
         protected ProductService $productService,
     ) {}
@@ -87,24 +89,5 @@ class ShopController extends Controller
             'relatedProducts' => $relatedProducts,
             'whatsappNumber' => $whatsappNumber,
         ]);
-    }
-
-    protected function formatWhatsAppNumber(string $number): string
-    {
-        $number = preg_replace('/[^0-9]/', '', $number);
-
-        if (str_starts_with($number, '08')) {
-            $number = '62' . substr($number, 1);
-        }
-
-        if (str_starts_with($number, '628')) {
-            return $number;
-        }
-
-        if (str_starts_with($number, '62')) {
-            return $number;
-        }
-
-        return '62' . $number;
     }
 }